Module zwp_virtual_keyboard::virtual_keyboard_unstable_v1::zwp_virtual_keyboard_v1

virtual keyboard

The virtual keyboard provides an application with requests which emulate the behaviour of a physical keyboard.

This interface can be used by clients on its own to provide raw input events, or it can accompany the input method protocol.

Structs

ZwpVirtualKeyboardV1

Enums

Error

Event

Request

Constants

REQ_DESTROY_SINCE
The minimal object version supporting this request

REQ_KEYMAP_SINCE
The minimal object version supporting this request

REQ_KEY_SINCE
The minimal object version supporting this request

REQ_MODIFIERS_SINCE
The minimal object version supporting this request

Statics

zwp_virtual_keyboard_v1_interface
C representation of this interface, for interop

zwp_virtual_keyboard_v1_requests
C-representation of the messages of this interface, for interop